Create base classes for WebLookup and DemocracyClub exports
..and update existing import scripts to use them
