+++ polling_stations/apps/data_collection/management/commands/__init__.py
@@ -10,6 +10,7 @@ from data_collection.base_importers import (
     BaseApiCsvStationsShpZipDistrictsImporter,
 )
 from data_collection.ems_importers import (
-    BaseXpressCsvImporter,
+    BaseXpressWebLookupCsvImporter,
+    BaseXpressDemocracyClubCsvImporter,
     BaseHalaroseCsvImporter,
 )
