+++ polling_stations/apps/data_collection/management/commands/import_cotswold.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id = 'E07000079'
     addresses_name = 'CotswoldsPropertyPostCodePollingStationWebLookup-2017-02-28.TSV'
     stations_name = 'CotswoldsPropertyPostCodePollingStationWebLookup-2017-02-28.TSV'
