+++ polling_stations/apps/data_collection/management/commands/import_isle_of_wight.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id      = 'E06000046'
     addresses_name  = 'rev02-2017/IsleofWightPropertyPostCodePollingStationWebLookup-2017-02-07.TSV'
     stations_name   = 'rev02-2017/IsleofWightPropertyPostCodePollingStationWebLookup-2017-02-07.TSV'
