+++ polling_stations/apps/data_collection/management/commands/import_malvern_hills.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id      = 'E07000235'
     addresses_name  = 'rev02-2017/MalvernHillsPropertyPostCodePollingStationWebLookup-2017-02-03.CSV'
     stations_name   = 'rev02-2017/MalvernHillsPropertyPostCodePollingStationWebLookup-2017-02-03.CSV'
