+++ polling_stations/apps/data_collection/management/commands/import_oldham.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id      = 'E08000004'
     addresses_name  = '21 Mar new version/Oldham2PropertyPostCodePollingStationWebLookup-2017-03-20.TSV'
     stations_name   = '21 Mar new version/Oldham2PropertyPostCodePollingStationWebLookup-2017-03-20.TSV'
