+++ polling_stations/apps/data_collection/management/commands/import_pendle.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id = 'E07000122'
     addresses_name = 'PendlePropertyPostCodePollingStationWebLookup-2017-03-01.TSV'
     stations_name = 'PendlePropertyPostCodePollingStationWebLookup-2017-03-01.TSV'
