+++ polling_stations/apps/data_collection/management/commands/import_preston.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id      = 'E07000123'
     addresses_name  = 'PrestonPropertyPostCodePollingStationWebLookup-2017-02-08.TSV'
     stations_name   = 'PrestonPropertyPostCodePollingStationWebLookup-2017-02-08.TSV'
