+++ polling_stations/apps/data_collection/management/commands/import_redditch.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id      = 'E07000236'
     addresses_name  = 'RBC Polling Station look up data.csv'
     stations_name   = 'RBC Polling Station look up data.csv'
