+++ polling_stations/apps/data_collection/management/commands/import_stockport.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id      = 'E08000007'
     addresses_name  = 'StockportPropertyPostCodePollingStationWebLookup-2017-02-16.CSV'
     stations_name   = 'StockportPropertyPostCodePollingStationWebLookup-2017-02-16.CSV'
