+++ polling_stations/apps/data_collection/management/commands/import_trafford.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id      = 'E08000009'
     addresses_name  = 'TraffordPropertyPostCodePollingStationWebLookup-2017-02-10.csv'
     stations_name   = 'TraffordPropertyPostCodePollingStationWebLookup-2017-02-10.csv'
