+++ polling_stations/apps/data_collection/management/commands/import_wychavon.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id      = 'E07000238'
     addresses_name  = 'WychavonPropertyPostCodePollingStationWebLookup-2017-02-13.TSV'
     stations_name   = 'WychavonPropertyPostCodePollingStationWebLookup-2017-02-13.TSV'
