+++ polling_stations/apps/data_collection/management/commands/import_wyre_forest.py
@@ -1,6 +1,6 @@
-from data_collection.management.commands import BaseXpressCsvImporter
+from data_collection.management.commands import BaseXpressWebLookupCsvImporter
 
-class Command(BaseXpressCsvImporter):
+class Command(BaseXpressWebLookupCsvImporter):
     council_id      = 'E07000239'
     addresses_name  = 'WyreForestPropertyPostCodePollingStationWebLookup-2017-02-09.csv'
     stations_name   = 'WyreForestPropertyPostCodePollingStationWebLookup-2017-02-09.csv'
